requirements.txt
edge-tts>=6.1.0
moviepy>=1.0.3
Pillow>=10.0.0
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
openai>=1.10.0
python-dotenv>=1.0.0
